requirements.txt
pyautogui
opencv-python
pillow
requests
python-dotenv